Move namespaces/files
