+++ src/Controls/src/Core/Compatibility/Handlers/ListView/Android/EntryCellView.cs
@@ -85,7 +85,7 @@ namespace Microsoft.Maui.Controls.Handlers.Compatibility
 			{
 				OnKeyboardDoneButtonPressed(EditText, EventArgs.Empty);
 				EditText.ClearFocus();
-				v.HideKeyboard();
+				v.HideSoftInput();
 			}
 
 			// Fire Completed and dismiss keyboard for hardware / physical keyboards
@@ -93,7 +93,7 @@ namespace Microsoft.Maui.Controls.Handlers.Compatibility
 			{
 				OnKeyboardDoneButtonPressed(EditText, EventArgs.Empty);
 				EditText.ClearFocus();
-				v.HideKeyboard();
+				v.HideSoftInput();
 			}
 
 			return true;
