+++ src/Core/src/Platform/Windows/KeyboardExtensions.cs
@@ -1,13 +1,10 @@
 ﻿using System;
 using System.Diagnostics;
-using Microsoft.UI.Xaml;
-using System.Diagnostics.CodeAnalysis;
 using Microsoft.UI.Xaml.Input;
-using Windows.UI.ViewManagement;
 
 namespace Microsoft.Maui.Platform
 {
-	public static partial class KeyboardExtensions
+	public static class KeyboardExtensions
 	{
 		public static InputScopeName ToInputScopeName(this Keyboard self)
 		{
@@ -158,50 +155,6 @@ namespace Microsoft.Maui.Platform
 
 			result.Names.Add(name);
 			return result;
-		}
-
-		internal static bool HideKeyboard(this FrameworkElement element)
-		{
-			if (TryGetInputPane(out var inputPane))
-			{
-				return inputPane.TryHide();
-			}
-
-			return false;
-		}
-
-		internal static bool ShowKeyboard(this FrameworkElement element)
-		{
-			if (TryGetInputPane(out var inputPane))
-			{
-				return inputPane.TryShow();
-			}
-
-			return false;
-		}
-
-		internal static bool IsSoftKeyboardShowing(this FrameworkElement element)
-		{
-			if (TryGetInputPane(out var inputPane))
-			{
-				return inputPane.Visible;
-			}
-
-			return false;
-		}
-
-		internal static bool TryGetInputPane([NotNullWhen(true)] out InputPane? inputPane)
-		{
-			var handleId = Process.GetCurrentProcess().MainWindowHandle;
-			if (handleId == IntPtr.Zero)
-			{
-				inputPane = null;
-
-				return false;
-			}
-
-			inputPane = InputPaneInterop.GetForWindow(handleId);
-			return true;
 		}
 	}
 }
